+++ release.config.js
@@ -12,6 +12,8 @@ const configuration = {
 
 configuration.plugins.push(['@semantic-release/commit-analyzer', {
     'releaseRules': [
+        {'type': 'chore', 'scope': 'deps', 'release': 'patch'},
+        {'type': 'chore', 'scope': 'package', 'release': 'patch'},
         {'type': 'build', 'scope': 'deps', 'release': 'patch'},
         {'type': 'docs', 'release': 'patch'}
     ]
@@ -42,19 +44,6 @@ configuration.plugins.push(['@semantic-release/git', {
     'assets': ['CHANGELOG.md', 'LICENSE'],
     'message': 'chore(release): :bookmark: ${nextRelease.version} [skip ci]\n\n${nextRelease.notes}'
 }]);
-
-const dockerImages = [];
-if (process.env.DOCKER_LOCAL_IMAGE_DH) {
-    dockerImages.push(process.env.DOCKER_LOCAL_IMAGE_DH);
-}
-if (process.env.DOCKER_LOCAL_IMAGE_GH) {
-    dockerImages.push(process.env.DOCKER_LOCAL_IMAGE_GH);
-}
-if(dockerImages.length > 0) {
-    configuration.plugins.push(['@sebbo2002/semantic-release-docker', {
-        images: dockerImages
-    }]);
-}
 
 configuration.plugins.push(['@qiwi/semantic-release-gh-pages-plugin', {
     'msg': 'docs: Updated for <%= nextRelease.gitTag %>',