+++ test/event.ts
@@ -36,6 +36,9 @@ describe('ical-generator Event', function () {
                 busystatus: ICalEventBusyStatus.BUSY,
                 priority: 5,
                 url: 'https://github.com/sebbo2002/ical-generator',
+                attachments: [
+                    'https://files.sebbo.net/calendar/attachments/foo'
+                ],
                 transparency: ICalEventTransparency.TRANSPARENT,
                 created: new Date().toJSON(),
                 lastModified: new Date().toJSON(),
@@ -1206,14 +1209,31 @@ describe('ical-generator Event', function () {
             assert.deepStrictEqual(event.organizer(), {
                 name: 'Sebastian Pekarek',
                 email: 'mail@example.com',
-                mailto: undefined
+                mailto: undefined,
+                sentBy: undefined
             });
 
             event.organizer({name: 'Sebastian Pekarek', email: 'mail@example.com', mailto: 'mail2@example2.com'});
             assert.deepStrictEqual(event.organizer(), {
                 name: 'Sebastian Pekarek',
                 email: 'mail@example.com',
-                mailto: 'mail2@example2.com'
+                mailto: 'mail2@example2.com',
+                sentBy: undefined
+            });
+        });
+
+        it('should support sent by when using object', function () {
+            const event = new ICalEvent({
+                start: moment(),
+                summary: 'Example Event'
+            }, new ICalCalendar());
+
+            event.organizer({name: 'Sebastian Pekarek', email: 'mail@example.com', sentBy: 'bot@example.com'});
+            assert.deepStrictEqual(event.organizer(), {
+                name: 'Sebastian Pekarek',
+                email: 'mail@example.com',
+                mailto: undefined,
+                sentBy: 'bot@example.com'
             });
         });
 
@@ -1264,7 +1284,8 @@ describe('ical-generator Event', function () {
             assert.deepStrictEqual(event.organizer(), {
                 name: 'Sebastian Pekarek',
                 email: undefined,
-                mailto: undefined
+                mailto: undefined,
+                sentBy: undefined
             });
         });
     });
@@ -1558,6 +1579,35 @@ describe('ical-generator Event', function () {
 
             event.url('http://github.com/sebbo2002/ical-generator');
             assert.strictEqual(event.url(), 'http://github.com/sebbo2002/ical-generator');
+        });
+    });
+
+    describe('createAttachment()', function () {
+        it('should return this', function () {
+            const event = new ICalEvent({}, new ICalCalendar());
+            assert.deepStrictEqual(event.createAttachment('https://files.sebbo.net/calendar/attachments/foo'), event);
+        });
+    });
+
+    describe('attachments()', function () {
+        it('getter should return an array of strings…', function () {
+            const event = new ICalEvent({}, new ICalCalendar());
+            assert.strictEqual(event.attachments().length, 0);
+
+            event.createAttachment('https://files.sebbo.net/calendar/attachments/foo');
+            assert.strictEqual(event.attachments().length, 1);
+            assert.deepStrictEqual(typeof event.attachments()[0], 'string');
+        });
+
+        it('setter should add url and return this', function () {
+            const event = new ICalEvent({}, new ICalCalendar());
+            const foo = event.attachments([
+                'https://files.sebbo.net/calendar/attachments/foo',
+                'https://files.sebbo.net/calendar/attachments/bar'
+            ]);
+
+            assert.strictEqual(event.attachments().length, 2);
+            assert.deepStrictEqual(foo, event);
         });
     });