+++ test/tools.ts
@@ -1,7 +1,8 @@
 'use strict';
 
 import assert from 'assert';
-import moment from 'moment-timezone';
+import moment from 'moment';
+import momentTz from 'moment-timezone';
 import {DateTime} from 'luxon';
 import dayjs from 'dayjs';
 import dayJsUTCPlugin from 'dayjs/plugin/utc';
@@ -79,6 +80,26 @@ describe('ICalTools', function () {
             it('should work with floating flag', function () {
                 assert.strictEqual(
                     formatDate(null, moment('2018-07-05T18:24:00.052'), false, true),
+                    '20180705T182400'
+                );
+            });
+        });
+        describe('moment-timezone.js', function () {
+            it('should work without setting a timezone', function () {
+                assert.strictEqual(
+                    formatDate(null, momentTz('2018-07-05T18:24:00.052Z'), false, false),
+                    '20180705T182400Z'
+                );
+            });
+            it('should work with timezone in event / calendar (with moment-timezone)', function () {
+                assert.strictEqual(
+                    formatDate('Canada/Saskatchewan', momentTz('2018-07-05T18:24:00.052Z'), false, false),
+                    '20180705T122400'
+                );
+            });
+            it('should work with floating flag', function () {
+                assert.strictEqual(
+                    formatDate(null, momentTz('2018-07-05T18:24:00.052'), false, true),
                     '20180705T182400'
                 );
             });
@@ -162,38 +183,50 @@ describe('ICalTools', function () {
     describe('escape()', function () {
         it('should escape \\', function () {
             assert.strictEqual(
-                escape('Lorem \\ipsum'),
+                escape('Lorem \\ipsum', false),
                 'Lorem \\\\ipsum'
             );
         });
         it('should escape ;', function () {
             assert.strictEqual(
-                escape('Lorem ;ipsum'),
+                escape('Lorem ;ipsum', false),
                 'Lorem \\;ipsum'
             );
         });
         it('should escape ,', function () {
             assert.strictEqual(
-                escape('Lorem, ipsum'),
+                escape('Lorem, ipsum', false),
                 'Lorem\\, ipsum'
             );
         });
         it('should escape \\r', function () {
             assert.strictEqual(
-                escape('Lorem \ripsum'),
+                escape('Lorem \ripsum', false),
                 'Lorem \\nipsum'
             );
         });
         it('should escape \\n', function () {
             assert.strictEqual(
-                escape('Lorem \nipsum'),
+                escape('Lorem \nipsum', false),
                 'Lorem \\nipsum'
             );
         });
         it('should escape \\r\\n', function () {
             assert.strictEqual(
-                escape('Lorem \r\nipsum'),
+                escape('Lorem \r\nipsum', false),
                 'Lorem \\nipsum'
+            );
+        });
+        it('should escape " in text when inQuotes = true', function () {
+            assert.strictEqual(
+                escape('Lorem "ipsum', true),
+                'Lorem \\"ipsum'
+            );
+        });
+        it('should not escape " in text when inQuotes = false', function () {
+            assert.strictEqual(
+                escape('Lorem "ipsum', false),
+                'Lorem "ipsum'
             );
         });
     });
@@ -288,6 +321,10 @@ describe('ICalTools', function () {
         it('should work with moment object', function () {
             const date = new Date();
             assert.deepStrictEqual(toDate(moment(date)), date);
+        });
+        it('should work with moment-timezone object', function () {
+            const date = new Date();
+            assert.deepStrictEqual(toDate(momentTz(date)), date);
         });
         it('should work with Day.js object', function () {
             const date = new Date();